mcp-servers/tee-web-agent/requirements.txt
browser-use>=0.1.40
playwright>=1.40.0
langchain-openai>=0.2.0
langchain-anthropic>=0.3.0
langchain>=0.3.0
eth-account>=0.10.0
web3>=6.14.0
eciespy>=0.4.0
httpx>=0.27.0
python-dotenv>=1.0.0
hedera-sdk-python>=2.0.0
mcp>=1.0.0
